+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Add accessors for the default options applied to newly-spawned tracees
+
 ### Changed
 
 - Update MSRV to 1.64